+++ src/pages/ModelizerDrawPage.vue
@@ -16,7 +16,7 @@ import ComponentDropOverlay from 'components/drawer/ComponentDropOverlay';
 import {
   addNewComponent,
   addNewTemplateComponent,
-  getPluginByName,
+  getPluginByName, initComponents,
   renderConfiguration,
   renderModel,
 } from 'src/composables/PluginManager';
@@ -32,6 +32,7 @@ import { useRoute } from 'vue-router';
 import { Notify } from 'quasar';
 import { useI18n } from 'vue-i18n';
 import PluginEvent from 'src/composables/events/PluginEvent';
+import { getTemplatesByType } from 'src/composables/TemplateManager';
 
 const { t } = useI18n();
 const route = useRoute();
@@ -43,9 +44,6 @@ const data = reactive({
   plugin: null,
 });
 const templates = ref([]);
-const defaultFolder = ref(process.env.MODELS_DEFAULT_FOLDER !== ''
-  ? `${process.env.MODELS_DEFAULT_FOLDER}/`
-  : '');
 
 let pluginDefaultSubscription;
 let pluginInitSubscription;
@@ -64,17 +62,55 @@ async function onDefaultEvent({ event }) {
     if (event.action === 'move') {
       await renderConfiguration(
         projectName.value,
-        `${defaultFolder.value}${query.value.path}`,
+        query.value.path,
         data.plugin,
       );
     } else if (renderModelActions.includes(event.action)) {
       await renderModel(
         projectName.value,
-        `${defaultFolder.value}${query.value.path}`,
+        query.value.path,
         data.plugin,
       );
     }
   }
+}
+
+/**
+ * Update plugin, draw components and update component templates array.
+ * @return {Promise<void>} Promise with nothing on success otherwise an error.
+ */
+async function initView() {
+  if (!query.value?.path) {
+    return;
+  }
+
+  data.plugin = getPluginByName(query.value.plugin);
+
+  if (!data.plugin) {
+    return;
+  }
+
+  await Promise.allSettled([
+    initComponents(
+      route.params.projectName,
+      data.plugin,
+      `${query.value.path}`,
+    ).then(() => {
+      data.plugin.draw('root');
+    }),
+    getTemplatesByType(
+      'component',
+      data.plugin.data.name,
+    ).then((response) => {
+      templates.value = response;
+    }).catch(() => {
+      Notify.create({
+        type: 'negative',
+        message: t('errors.templates.getData'),
+        html: true,
+      });
+    }),
+  ]);
 }
 
 /**
@@ -110,7 +146,7 @@ async function dropHandler(event) {
     await addNewComponent(
       route.params.projectName,
       data.plugin,
-      `${defaultFolder.value}${route.query.path}`,
+      query.value.path,
       componentDefinition,
       new ComponentDrawOption({ ...getComponentPosition(event) }),
     );
@@ -123,7 +159,7 @@ async function dropHandler(event) {
     await addNewTemplateComponent(
       route.params.projectName,
       data.plugin,
-      `${defaultFolder.value}${route.query.path}`,
+      query.value.path,
       templateDefinition,
     ).then(() => {
       data.plugin.draw('root');
@@ -135,6 +171,11 @@ async function dropHandler(event) {
       });
     });
   }
+  await renderModel(
+    projectName.value,
+    query.value.path,
+    data.plugin,
+  );
 }
 
 onMounted(async () => {
@@ -142,9 +183,9 @@ onMounted(async () => {
     onDefaultEvent(event);
   });
   pluginInitSubscription = PluginEvent.InitEvent.subscribe(() => {
-    data.plugin = getPluginByName(query.value.path.split('/')[0]);
+    initView();
   });
-  data.plugin = getPluginByName(query.value.path.split('/')[0]);
+  await initView();
 });
 
 onUnmounted(() => {